requirements.txt
numpy
torch
Pillow
scikit-image
pytest